Update index.html with header, intro, and topsells
In the index.html file, a header is added to the top for navigation
links. Below that is an intro to the website and the purpose of
its existence. Beneath that is a "best-sellers" category for advertising
very popular gaming consoles that are currently getting a lot of
attraction on the website.